+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to `laravel-nano-jsx-preset` will be documented in this file
 
+## v0.1.1 - 2021-12-28
+
+- Added `<Helmet />` example in `App.tsx`
+- Added dynamic title example in `App.tsx`
+- Added 404 Error example route in `App.tsx`
+- Added `exact` to all routes in `App.tsx`
+- Added wildcard `path` to `Create routes` description (fixes routes)
+- Added `<small />` css attribute
+- Added .disableNotifications() to webpack.mix.js
+
+- Replaced global router import with explicit imports
+
 ## v0.1.0 - 2021-12-07
 
 - Initial release